start making player mvc
